# public/api/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.0",
        "title": "fs2netd REST API"
    },
    "basePath": "/api/v1",
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "user_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/users": {
            "get": {
                "description": "Returns a list of all registered users\n\n\n**Note:** The user referenced by the token needs to have administrator rights to execute this function\n",
                "security": [
                    {
                        "user_token": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesfull request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "400": {
                        "description": "Malformed input",
                        "schema": {
                            "$ref": "#/definitions/InputError"
                        }
                    },
                    "401": {
                        "description": "No or invalid API token specified",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The user of the token has insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "A server error occured",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Registers a new user.\n",
                "responses": {
                    "201": {
                        "description": "Succesfully registered the user",
                        "schema": {
                            "$ref": "#/definitions/RegisteredUser"
                        }
                    },
                    "409": {
                        "description": "The specified username is already used",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "A server error occured",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "description": "Retrieves extended information about the user of the used API token\n\n\n**Note:** A valid API token is required for this function\n",
                "security": [
                    {
                        "user_token": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved information",
                        "schema": {
                            "$ref": "#/definitions/ExtendedUser"
                        }
                    },
                    "400": {
                        "description": "Malformed input",
                        "schema": {
                            "$ref": "#/definitions/InputError"
                        }
                    },
                    "401": {
                        "description": "No or invalid API token specified",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The user of the token has insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "A server error occured",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "The resource id",
                    "required": true,
                    "type": "integer"
                }
            ],
            "get": {
                "description": "Retrieves extended information about a user\n\n\n**Note:** The user referenced by the token needs to have administrator rights to execute this function\n",
                "security": [
                    {
                        "user_token": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved information",
                        "schema": {
                            "$ref": "#/definitions/ExtendedUser"
                        }
                    },
                    "400": {
                        "description": "Malformed input",
                        "schema": {
                            "$ref": "#/definitions/InputError"
                        }
                    },
                    "401": {
                        "description": "No or invalid API token specified",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The user of the token has insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "A server error occured",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a user from the server\n\n\n**Note:** The user referenced by the token needs to have administrator rights to execute this function\n",
                "security": [
                    {
                        "user_token": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the resource"
                    },
                    "400": {
                        "description": "Malformed input",
                        "schema": {
                            "$ref": "#/definitions/InputError"
                        }
                    },
                    "401": {
                        "description": "No or invalid API token specified",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The user of the token has insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "A server error occured",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "RegisteredUser": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "last_login": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "ExtendedUser": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "last_login": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                }
            }
        },
        "InputError": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}